Move image-data into data directory
