--- src/assets/images/image-data.ts
@@ -1,21 +1,0 @@
-import cholesterolImage from "./heart/cholesterol.png";
-import foamCellImage from "./heart/foam-cell.png";
-import immuneCellImage from "./heart/immune-cell.png";
-import plaqueImage from "./heart/plaque.png";
-import redBloodCellImage from "./heart/red-blood-cell.png";
-import smoothMuscleCellsImage from "./heart/smooth-muscle-cells.png";
-import stickyProteinImage from "./heart/sticky-protein.png";
-import vesselCellsImage from "./heart/vessel-cells.png";
-
-export const keyImages = {
-  "h": [
-    redBloodCellImage,
-    immuneCellImage,
-    cholesterolImage,
-    stickyProteinImage,
-    vesselCellsImage,
-    foamCellImage,
-    plaqueImage,
-    smoothMuscleCellsImage
-  ]
-};
