+++ src/data/image-data.ts
@@ -1,0 +1,21 @@
+import cholesterolImage from "../assets/images/heart/cholesterol.png";
+import foamCellImage from "../assets/images/heart/foam-cell.png";
+import immuneCellImage from "../assets/images/heart/immune-cell.png";
+import plaqueImage from "../assets/images/heart/plaque.png";
+import redBloodCellImage from "../assets/images/heart/red-blood-cell.png";
+import smoothMuscleCellsImage from "../assets/images/heart/smooth-muscle-cells.png";
+import stickyProteinImage from "../assets/images/heart/sticky-protein.png";
+import vesselCellsImage from "../assets/images/heart/vessel-cells.png";
+
+export const keyImages = {
+  "h": [
+    redBloodCellImage,
+    immuneCellImage,
+    cholesterolImage,
+    stickyProteinImage,
+    vesselCellsImage,
+    foamCellImage,
+    plaqueImage,
+    smoothMuscleCellsImage
+  ]
+};
